Quiet core/test "close multiple times" test
As folks may know better than I, in uTest individual tests are documented
are documented as returning a value, which can itself be validated.
If the return value is not Unit, it is printed out.

Before this PR, the "close multiple times" test was returning the
value of an object, which was getting printed out and adding
something to the log which needed to be figured out as safe.

This PR causes that test to be silent, thereby reducing clutter.

Agreed, this is a nit. I think it is worth fixing because the present
file may serve as a cut & paste template for additional tests as they
are added.

If I can not fix something useful, I can at least remove a small wart.